feat: 게시물 수정 기능 추가

+++ src/pages/ProductInform.js
@@ -29,8 +29,15 @@ function ProductInform({match}) {
     }
     console.log(info)
     const deletePost = async () => {
-        await axios.delete(`http://localhost:8080/products/${info?.productID}`);
-        window.location.replace("/detail");
+      if(window.confirm('해당 게시물을 삭제하시겠습니까?\n삭제된 데이터는 복구할 수 없습니다.')) {
+          await axios.delete(`http://localhost:8080/products/${info?.productID}`);
+          alert('게시물이 삭제되었습니다.')
+          return window.location.href = "/detail";
+      }
+    }
+    const editPost = async () => {
+        localStorage.setItem("data", JSON.stringify(info));
+        return window.location.href = `/edit/${info.productID}`;
     }
   return (
       <div>
@@ -55,12 +62,20 @@ function ProductInform({match}) {
             </div>
             <div><Title>{info.title}</Title></div>
             <div><Content>{info.content}</Content></div>
-            <Btn>쪽지하기</Btn>
-            <Link to={`/detail`}><Btn>목록으로</Btn></Link>
-            {user === info.accountID
-                ?(<UserBtn style={{'background':'#8D192B'}} onClick={deletePost}>삭제</UserBtn>)
-                :''
-            }
+              {user === info.accountID
+                  ?(
+                      <div style={{'display': 'inline-flex', 'width':'700px', 'justify-content':'flex-end'}}>
+                          <UserBtn style={{'background':'#0080FF'}} onClick={editPost}>수정</UserBtn>
+                          <UserBtn style={{'background':'#8D192B'}} onClick={deletePost}>삭제</UserBtn>
+                      </div>
+                  )
+                  :''
+              }
+            <div>
+                <Btn>쪽지하기</Btn>
+                <Link to={`/detail`}><Btn>목록으로</Btn></Link>
+            </div>
+
           </div>
         ):''}
       </div>
@@ -144,10 +159,10 @@ background: #375945;
 color: #FFFFFF;
 `
 const UserBtn = styled.button`
-width: 5%; padding: 0.6%;
-margin: 2%;
+width: 8%; padding: 0.5%;
+margin: 1%;
 font-size: 80%;
 cursor: pointer;
-border: none; border-radius: 50px;
+border: none; border-radius: 5px;
 color: #FFFFFF;
 `
